Integrated ColorMixin into ColorAxisComposition.
